add logOut

+++ src/redux/contacts/operations.js
@@ -4,10 +4,16 @@ import axios from "../../operationsAPI";
 export const fetchContacts = createAsyncThunk(
   "contacts/fetchAll",
   async (_, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("User is not authenticated");
+    }
+
     try {
       const response = (await axios.get("/contacts")).data;
       return response;
-      // console.log(response);
     } catch (e) {
       return thunkAPI.rejectWithValue(e.message);
     }
@@ -17,6 +23,13 @@ export const fetchContacts = createAsyncThunk(
 export const addContact = createAsyncThunk(
   "contacts/addContact",
   async (contact, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("User is not authenticated");
+    }
+
     try {
       const response = (await axios.post("/contacts", contact)).data;
       return response;
@@ -29,6 +42,13 @@ export const addContact = createAsyncThunk(
 export const deleteContact = createAsyncThunk(
   "contacts/deleteContact",
   async (contactId, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("User is not authenticated");
+    }
+
     try {
       const response = (await axios.delete(`/contacts/${contactId}`)).data;
       return response;
@@ -41,6 +61,13 @@ export const deleteContact = createAsyncThunk(
 export const editContact = createAsyncThunk(
   "contacts/editContact",
   async (contact, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("User is not authenticated");
+    }
+
     try {
       const { id, ...updatedFields } = contact;
       const response = (await axios.patch(`/contacts/${id}`, updatedFields))
